Fixed plugin manager not opening for native-image
It tried to load an image that doesn't exist yet, so it raised an error and prevented the plugin manager from opening. The resource loading method for that image has been replaced with a safer solution that doesn't return null if it doesn't exist.

If you say I could've placed an image there to solve the issue, don't even mention it